Refactor DBConnection class
The class was heavily refactored, making connect and disconnect
private methods that are called by the run_query method. This means
that all you need to do with the class is run_query and it will
do the rest.

There is an outstanding issue with this approach though. I cannot
work out how to test for the case when connecting to the database
fails. The way I've made the connect method private means that
it is not visible to the RSpec allow_any_instance_of method, which
is required because the DBConnection methods are class methods.